[SC-65]  Refactoring : Like, Scrap Post 패키지 내부로 이동
