+++ src/main/java/inu/codin/codin/domain/post/like/LikeEntity.java
@@ -1,4 +1,4 @@
-package inu.codin.codin.domain.scrap;
+package inu.codin.codin.domain.post.like;
 
 import inu.codin.codin.common.BaseTimeEntity;
 import lombok.Builder;
@@ -6,16 +6,16 @@ import lombok.Getter;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-@Document(collection = "scraps")
+@Document(collection = "likes")
 @Getter
-public class ScrapEntity extends BaseTimeEntity {
+public class LikeEntity extends BaseTimeEntity {
     @Id
     private String id;
     private String postId;
     private String userId;
 
     @Builder
-    public ScrapEntity(String postId, String userId) {
+    public LikeEntity(String postId, String userId) {
         this.postId = postId;
         this.userId = userId;
     }
