+++ src/main/java/inu/codin/codin/domain/post/scrap/ScrapEntity.java
@@ -1,4 +1,4 @@
-package inu.codin.codin.domain.like;
+package inu.codin.codin.domain.post.scrap;
 
 import inu.codin.codin.common.BaseTimeEntity;
 import lombok.Builder;
@@ -6,16 +6,16 @@ import lombok.Getter;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-@Document(collection = "likes")
+@Document(collection = "scraps")
 @Getter
-public class LikeEntity extends BaseTimeEntity {
+public class ScrapEntity extends BaseTimeEntity {
     @Id
     private String id;
     private String postId;
     private String userId;
 
     @Builder
-    public LikeEntity(String postId, String userId) {
+    public ScrapEntity(String postId, String userId) {
         this.postId = postId;
         this.userId = userId;
     }
