+++ src/main/java/inu/codin/codin/domain/post/scrap/ScrapRepository.java
@@ -1,9 +1,9 @@
-package inu.codin.codin.domain.like;
+package inu.codin.codin.domain.post.scrap;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LikeRepository extends MongoRepository<LikeEntity, String> {
+public interface ScrapRepository extends MongoRepository<ScrapEntity, String> {
     boolean findByPostIdAndUserId(String postId, String userId);
 }
